requirements.txt
astroid==2.3.3
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pkg-resources==0.0.0
pylint==2.4.4
pyserial==3.4
python-dotenv==0.13.0
rope==0.16.0
six==1.14.0
wrapt==1.11.2
